[MV] beesdoo_shift, beesdoo_worker_status: move constraint to specific module
Constraint on shift type should be move to specific module for beescoop
workflow
